Added support for multiple commands ala git
To define new command just inherit from SS2JSON::Command

This is an example

    require 'ss2json/command'

    class ToHash < SS2JSON::Command

      def self.name
        "to_hash"
      end

      def self.description
        "Convert a spreadsheet to a json hash"
      end

      def initialize(options)
        puts options.inspect
      end

    end

From the command line point of view, it acts as git.
All the commands should be stored in 'ss2json-commands/command_name.rb'
and the name should be equal to the file name. The commands could be in
different rubygems.